add debug log for
Took 33 minutes

+++ src/main/java/de/staticred/dbv2/discord/events/MessageEvent.java
@@ -2,6 +2,7 @@ package de.staticred.dbv2.discord.events;
 
 import de.staticred.dbv2.DBUtil;
 import de.staticred.dbv2.files.filehandlers.ConfigFileManager;
+import de.staticred.dbv2.util.Logger;
 import net.dv8tion.jda.api.events.message.guild.GuildMessageReceivedEvent;
 import net.dv8tion.jda.api.hooks.ListenerAdapter;
 
@@ -30,16 +31,25 @@ public class MessageEvent extends ListenerAdapter {
         if (event.getMember().getUser().isBot())
             return;
 
+        if (!DBUtil.getINSTANCE().getConfigFileManager().enabledDiscordMessages()) {
+            return;
+        }
+
         ConfigFileManager config = DBUtil.getINSTANCE().getConfigFileManager();
+
+        Logger debug = DBUtil.getINSTANCE().getLogger();
 
         int deleteTime = config.deleteTime();
 
         if (config.forceCleanChannel()) {
+            debug.postDebug("Clean channel enabled");
             List<String> channelIDs = config.getChanelIDs();
 
             String channelID = event.getChannel().getId();
 
             if (channelIDs.isEmpty()) {
+                debug.postDebug("CCList is empty");
+                debug.postDebug("Executing command handler");
                 DBUtil.getINSTANCE().getCommandManager().handleDiscordInput(event.getMember(), event.getChannel(), event.getMessage(), event.getMessage().getContentRaw());
                 if (deleteTime > -1)
                     event.getMessage().delete().queueAfter(deleteTime, TimeUnit.SECONDS);
@@ -47,15 +57,18 @@ public class MessageEvent extends ListenerAdapter {
             }
 
             if (!channelIDs.contains(channelID)) {
+                debug.postDebug("Wrong channel detected, ignoring messages.");
                 return;
             }
 
             if (!config.removeOwnerMessages() && event.getMember().isOwner()) {
+                debug.postDebug("Executing command handler");
                 DBUtil.getINSTANCE().getCommandManager().handleDiscordInput(event.getMember(), event.getChannel(), event.getMessage(), event.getMessage().getContentRaw());
                 return;
             }
 
             if (!DBUtil.getINSTANCE().getCommandManager().handleDiscordInput(event.getMember(), event.getChannel(), event.getMessage(), event.getMessage().getContentRaw())) {
+                debug.postDebug("Invalid command input, deleting message");
                 event.getMessage().delete().queue();
                 return;
             }
@@ -63,6 +76,7 @@ public class MessageEvent extends ListenerAdapter {
             if (deleteTime > -1)
                 event.getMessage().delete().queueAfter(deleteTime, TimeUnit.SECONDS);
         } else {
+            debug.postDebug("Executing command handler");
             if (DBUtil.getINSTANCE().getCommandManager().handleDiscordInput(event.getMember(), event.getChannel(), event.getMessage(), event.getMessage().getContentRaw())) {
                 if (deleteTime > -1) {
                     event.getMessage().delete().queueAfter(deleteTime, TimeUnit.SECONDS);
